improve phone dict to use U, WQ
phone_dict did not make good use of all phoneme shapes. In particular: 
Changed all of the Uxx phones to use U phoneme, not O. 
Changed W phone to use the WQ phoneme, not 'etc'. 
Changed AWx to use WQ instead of AI, I could be wrong, be nice if it used both.